requirements.txt
altair==5.4.0
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b3
azure-core==1.30.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
duckdb==1.1.1
eval==0.0.5
exceptiongroup==1.2.2
executing==2.0.1
Faker==19.13.0
fonttools==4.53.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.2
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mysql-connector-python==9.0.0
narwhals==1.4.1
numpy==1.26.4
openai==1.40.6
packaging==24.1
pandas==1.5.3
pandasai==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.27.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==1.10.17
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.53
sqlglot==25.24.1
sqlglotrs==0.2.12
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.39.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.20.0
